Revise error message sentence
A message about a weak password can be understood as a warning. Adding a "Error" prefix will help to understand the situation.

Reading "Failed to start" let the user think something wrong occurred, after trying to start the service. So, we look for other logs.
Changing vocabulary will help to understand that the script refused to start the service.